mv client auth functions from client.app to supernode.app
